+++ bin/console
@@ -7,8 +7,8 @@ require "lpw/statistics"
 # with your gem easier. You can also use a different console, if you like.
 
 # (If you use this, don't forget to add pry to your Gemfile!)
-# require "pry"
-# Pry.start
+require "pry"
+Pry.start
 
-require "irb"
-IRB.start
+# require "irb"
+# IRB.start
