+++ lib/lpw/statistics/version.rb
@@ -1,5 +1,5 @@
 module Lpw
   module Statistics
-    VERSION = "0.4.8"
+    VERSION = "1.0.0"
   end
 end
