+++ lpw-statistics.gemspec
@@ -33,4 +33,6 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "bundler", "~> 1.12"
   spec.add_development_dependency "rake", "~> 10.0"
+  spec.add_development_dependency "pry-coolline"
+  spec.add_development_dependency "pry"
 end
